Shapes in namespace because windows.h is gay
