+++ src/game.cpp
@@ -16,6 +16,9 @@ inline void loadStage(Game& game) {
 	if (loadStageFromFile("assets/levels/level1.json", game.stage, e)) {
 		game.stage.state.type = StageState::RUNNING;
 		createAABB(game.stage, AABB(game.stage.sea)); // TODO: make seas a list so there can be seas anywhere
+		return;
+	} else {
+		throw e;
 	}
 }
