+++ src/game.cxx
@@ -5,9 +5,11 @@
 #include "printing.hpp"
 #include "graphics.hpp"
 #include "serialize.hpp"
+#include "win32_file.hpp"
 
 
 int main() {
+	std::cout << ExePath() << std::endl;
 	sf::RenderWindow window;
 	sf::ContextSettings settings;
 	// sf::VideoMode videoMode = sf::VideoMode::getDesktopMode();
@@ -20,28 +22,33 @@ int main() {
 	 	std::cout << "Couldn't load font" << std::endl;
 	}
 
-	Clock drawClock;
-	Game game;
-	start(game);
-	while(!game.end && window.isOpen()) {
-		sf::Event e;
-		while(window.pollEvent(e)) {
-			processEvent(game, e, window);
-			if(e.type == sf::Event::Resized) {
-				auto view = window.getView();
-				view.setCenter({e.size.width/2.f, e.size.height/2.f});
-				view.setSize({(float)e.size.width, (float)e.size.height});
-				window.setView(view);
+	try {
+		Clock drawClock;
+		Game game;
+		start(game);
+		while(!game.end && window.isOpen()) {
+			sf::Event e;
+			while(window.pollEvent(e)) {
+				processEvent(game, e, window);
+				if(e.type == sf::Event::Resized) {
+					auto view = window.getView();
+					view.setCenter({e.size.width/2.f, e.size.height/2.f});
+					view.setSize({(float)e.size.width, (float)e.size.height});
+					window.setView(view);
+				}
+			}
+			update(game);
+			float drawDelta = drawClock.getElapsedTime().asSeconds();
+			if (drawDelta >= FRAME_RATE) {
+				drawStage(window, game.stage, true);
+				drawInfoText(window, game.stage, drawDelta, game.updateDelta, 0);
+				window.display();
+				drawClock.restart();
 			}
 		}
-		update(game);
-		float drawDelta = drawClock.getElapsedTime().asSeconds();
-		if (drawDelta >= FRAME_RATE) {
-			drawStage(window, game.stage, true);
-			drawInfoText(window, game.stage, drawDelta, game.updateDelta, 0);
-			window.display();
-			drawClock.restart();
-		}
+		stop(game);
+	} catch (SerializeError & e) {
+		std::cout << e.what << std::endl;
+		return 1;
 	}
-	stop(game);
 }
