+++ src/maths.hpp
@@ -4,6 +4,8 @@
 #include <cmath>
 #include <vector>
 
+#include "vector2.hpp"
+
 /********
  * Misc *
  ********/
@@ -18,105 +20,6 @@ T signOf(T t) {
 	return sign<T>(t);
 }
 
-
-/***********
- * Vectors *
- ***********/
-
-struct Vector2{
-	float& operator[](int i) { return (&x)[i]; } // read
-	const float& operator[](int i) const { return (&x)[i]; }; // write
-	float x = 0.f, y = 0.f;
-};
-
-const Vector2 VECTOR2_UP {0.f, 1.f};
-const Vector2 VECTOR2_DOWN {0.f, -1.f};
-const Vector2 VECTOR2_LEFT {-1.f, 0.f};
-const Vector2 VECTOR2_RIGHT {1.f, 0.f};
-const Vector2 VECTOR2_ZERO {0.f, 0.f};
-
-// Vector negation
-inline Vector2 operator-(const Vector2& v);
-// Vector/Vector addition and subtraction
-inline Vector2 operator+(const Vector2& b, const Vector2& a);
-inline Vector2 operator-(const Vector2& a, const Vector2& b);
-inline Vector2& operator+=(Vector2& a, const Vector2& b);
-inline Vector2& operator-=(Vector2& a, const Vector2& b);
-// Vector equality an inequality
-inline bool operator==(const Vector2& a, const Vector2& b);
-inline bool operator!=(const Vector2& a, const Vector2& b);
-// Vector/Scalar multiplication and addition
-inline Vector2 operator*(const Vector2& v, float s);
-inline Vector2 operator*(float s, const Vector2& v);
-inline Vector2 operator/(const Vector2& v, float s);
-inline Vector2 operator/(float s, const Vector2& v);
-inline Vector2& operator*=(Vector2& v, float s);
-inline Vector2& operator/=(Vector2& v, float s);
-// Vector math operations
-inline float dot(const Vector2& a, const Vector2& b);
-inline float cross(const Vector2& a, const Vector2& b);
-inline float squaredMagnitude(const Vector2& v);
-inline float magnitude(const Vector2& v);
-// normalize copy
-inline Vector2 normalized(const Vector2& v);
-// normalize in place
-inline Vector2& normalize(Vector2& v);
-inline Vector2& clamp(Vector2& v, float s); // clamp this vector to a scalar magnitude
-// returns + if point on left - if point on right 0 if point on line from b to a
-inline float sideSign(Vector2 a, Vector2 b, Vector2 point);
-// returns true if point is bounded on the line segment defined by a and b
-inline bool bounded(Vector2 a, Vector2 b, Vector2 point);
-
-#include "vector2.inl" // definitions for template functions and inlines
-
-/**********
- * Shapes *
- **********/
-
-struct Circle {
-	Vector2 position;
-	float radius = 0.f;
-};
-
-// NOTE: Polygon vertices in the array in order from top left to topleft clockwise
-// example:
-//      index 0					index 1
-//      	*---------------------*
-//      	|                     |
-//      	*---------------------*
-//      index 3					 index 2
-template <int N>
-struct Polygon {
-	Vector2 position;
-	Vector2 vertices[N];
-	Vector2 model[N];
-	int size = N;
-	float rotation = 0.f;
-};
-
-struct Rectangle : Polygon<4>{
-	float width = 0.f;
-	float height = 0.f;
-};
-
-inline float area(const Circle & circle);
-inline float area(Rectangle& rectangle);
-template <int N>
-inline void updateVertices(Polygon<N>& polygon);
-template <int N>
-inline bool pointInsidePolygon(Vector2 point, Polygon<N>& polygon);
-template <int N>
-inline Vector2 lower(const Polygon<N>& polygon); // gets the bottom left point of a polygon in relation to x, y axis
-template <int N>
-inline Vector2 upper(const Polygon<N>& polygon); // gets the top right point of a polygon in raltion to x, y axis
-template <int N>
-inline void boundingPoints(const Polygon<N>& polygon, Vector2& lower, Vector2& upper);
-
-
-#include "shapes.inl" // definitions for template functions and inlines
-
-Rectangle makeRectangle(Vector2 p, float w, float h, float rotation=0.f);
-
 /************
  * Geometry *
  ************/
@@ -124,3 +27,30 @@ Rectangle makeRectangle(Vector2 p, float w, float h, float rotation=0.f);
 bool lineSegmentIntersection(Vector2 a, Vector2 b, Vector2 c, Vector2 d, Vector2& intersection);
 Vector2 findNormal(Vector2 a, Vector2 b, Vector2 c);
 
+// returns + if point on left - if point on right 0 if point on line from b to a
+inline float sideSign(Vector2 a, Vector2 b, Vector2 point) {
+	Vector2 u = b - a;
+	Vector2 v = point - a;
+	float product = cross(u, v);
+	return sign(product);
+}
+
+inline bool bounded(Vector2 a, Vector2 b, Vector2 point) {
+	Vector2 ba = b - a;
+	if (std::abs(ba.x) > std::abs(ba.y)) { // line more horizontally oriented
+		// we check that we are bounded on x
+		if (ba.x > 0) { // positive dx means b.x > a.x
+			return point.x >= a.x && point.x <=b.x;
+		} else { // negative dx means that b.x < a.x
+			return point.x >= b.x && point.x <=a.x;
+		}
+	} else { // line is more vertically oriented
+		// we check that we are bounded on y
+		if (ba.y > 0) { // positive dy means b.y > a.y
+			return point.y >= a.y && point.y <= b.y;
+		} else { // negative dy means that a.y > b.y
+			return point.y >= b.y && point.y <= a.y;
+		}
+	}
+	assert(false);
+}
