+++ src/platform.cpp
@@ -2,6 +2,8 @@
 #include "platform.hpp"
 #include "stage.hpp"
 
+using namespace wabi;
+
 uint8 createPlatform(Stage& stage, Vector2 position, float width, float height) {
 	Platform platform;
 	platform.shape = makeRectangle(position, width, height);
@@ -22,6 +24,6 @@ uint8 createPlatform(Stage& stage, Rectangle rect) {
 
 PlatformIt findPlatform(Stage& stage, uint8 platformId) {
 	PlatformIt platformIt = std::lower_bound(stage.platforms.begin(), stage.platforms.end(), platformId, [](Platform& p, uint8 id) -> bool { return p.id < id; });
-	assert(platformIt != stage.platforms.end()); // ... then this assertion will fail ...
+	assert(platformIt != stage.platforms.end());
 	return platformIt;
 }
