+++ src/serialize.hpp
@@ -3,13 +3,14 @@
 #include <string>
 #include <vector>
 #include <sstream>
+#include "shapes.hpp"
 #include "stage.hpp"
 #include "json.hpp"
 
 // Serialize
 std::string serialize(const Stage& stage);
 std::string serialize(const std::vector<Platform> platforms);
-std::string serialize(const Rectangle& rectangle);
+std::string serialize(const wabi::Rectangle& rectangle);
 std::string serialize(const Vector2 v);
 std::string serialize(const Win w);
